requirements.txt
openai
boto3
pydub
python-dotenv